+++ backend/open_recipes/api/recipes.py
@@ -14,6 +14,7 @@ import sqlalchemy
 from sqlalchemy.exc import IntegrityError, OperationalError, ProgrammingError, DataError, SQLAlchemyError, DBAPIError
 import uvicorn
 from pydantic import BaseModel, ValidationError
+from api.auth import get_current_user, TokenData
 
 router = APIRouter(
   prefix="/recipes",
